クラスをpublicにするためにProductクラス、Calculateクラスのファイルをそれぞれ作成。
