+++ src/chapter5/BaseClass51.java
@@ -3,54 +3,6 @@ package chapter5;
 import java.io.*;
 
 /**
- * Productクラス。商品名と金額、税率をフィールドにもつ。
- */
-class Product {
-	private String name;
-	private int price;
-	private double tax = 0.1;
-	
-	/**
-	 * 商品名を設定する。
-	 * @param name 商品名
-	 */
-	public void setName(String name) {
-		this.name = name;
-	}
-	
-	/**
-	 * 商品名を取得する。
-	 * @return　商品名
-	 */
-	public String getName() {
-		return name;
-	}
-	
-	/**
-	 * 金額を設定する。
-	 * @param price　金額
-	 */
-	public void setPrice(int price) {
-		this.price = price;
-	}
-	
-	/**
-	 * 金額を取得する。
-	 * @return　金額
-	 */
-	public int getPrice() {
-		return price;
-	}
-
-	/**
-	 * @return 金額に税率(10%)をかけた金額を返す。
-	 */
-	public int calcTotalPrice() {
-		return (int) ( price + (price * tax));
-	}
-}
-
-/**
  * 商品名、金額と消費税率をフィールドに持つ商品クラスを作成してください。
  * 商品名と金額を入力し、計算結果を出力してください。消費税率は一律10%です。
  */
